requirements.txt
trimesh>=4.0.0
cascadio>=0.0.17
pyrender>=0.1.45
numpy>=1.24.0
Pillow>=10.0.0
google-genai>=1.0.0
fal-client>=0.4.0
python-dotenv>=1.0.0
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9
httpx>=0.27.0
pytest>=7.0.0
requests>=2.31.0
